Example of DI configuration in yaml/requirements.txt
PyYAML==6.0.2
fire==0.7.0
pandas==2.2.3
